Move animation view into example
